fix: split public and internal utils
