+++ lib/EasyPost/Pickup.php
@@ -2,7 +2,7 @@
 
 namespace EasyPost;
 
-use EasyPost\Util\Util;
+use EasyPost\Util\InternalUtil;
 
 /**
  * @package EasyPost
@@ -38,7 +38,7 @@ class Pickup extends EasyPostObject
      */
     public function lowestRate($carriers = [], $services = [])
     {
-        $lowestRate = Util::getLowestObjectRate($this, $carriers, $services, 'pickup_rates');
+        $lowestRate = InternalUtil::getLowestObjectRate($this, $carriers, $services, 'pickup_rates');
 
         return $lowestRate;
     }
